Implement right hand
It is not strictly symmetrical, so can't just be mirrored.

We need to mirror the switch footprint, so we can keep the same switch
after mirroring the entire shape in the slicer.

I would have mirrored the shape, but apparently opencascade doesn't
make it easy.

This produces a mirrored right hand. Remember to mirror it in the slicer
before printing.