Special case col 0 for the right hand
It needs a deeper hole due to the stagger.
Applying deeper holes for the entire keyboard like we did before is too
much of a compromise.

This is less elegant in terms of code, but it minimizes the differences
between right and left hand and makes it more likely the wires can all
have an easy path through their holes.